fix index dtype i4

+++ tests/test_simple_join.py
@@ -17,8 +17,8 @@ def test_simple_join():
         [0.0, 0.0, 6.0],
     ], dtype='f4')
 
-    index1 = np.array([0, 1, 2])
-    index2 = np.array([0, 1, 2])
+    index1 = np.array([0, 1, 2], dtype='i4')
+    index2 = np.array([0, 1, 2], dtype='i4')
 
     vdata, idata = mesh_join.join([vertices1, vertices2], [index1, index2], stride=12)
 
@@ -50,8 +50,8 @@ def test_join_with_restart_index():
         [0.0, 0.0, 6.0],
     ], dtype='f4')
 
-    index1 = np.array([0, 1, 2, -1, 2, 3])
-    index2 = np.array([0, 1])
+    index1 = np.array([0, 1, 2, -1, 2, 3], dtype='i4')
+    index2 = np.array([0, 1], dtype='i4')
 
     vdata, idata = mesh_join.join([vertices1, vertices2], [index1, index2], stride=12, separator=True)
